refactor(shell): fold the review findings into the rail
One `identity()` derives both the display name and the avatar initials from
the same fields in the same order, so the circle can no longer disagree with
the name beside it. The two link groups share one class. `initialsFor` is no
longer exported — nothing outside the file ever wanted it.

Layouts import `AppShell` through `@/features/shell`, the cross-feature API,
rather than reaching past the barrel.

The comments now name each place the artboard and the codebase disagree —
the missing "Taken courses" route, the two rail colours with no token behind
them, the drawer keeping the rail's own metrics, and the mobile title — so
the divergence is recorded where the code is, not only in the PR.
